Create new todo creation form and append it to template
